python-backend/requirements.txt
# CareLink Python Backend - Dépendances
# =======================================

# Framework Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6  # Pour upload de fichiers

# OCR
easyocr==1.7.1
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78

# Traitement d'images
Pillow==10.1.0
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4

# Validation et parsing
pydantic==2.5.0

# Logging et utilitaires
python-dotenv==1.0.0

# Note: EasyOCR téléchargera automatiquement les modèles français (~200MB)
# au premier lancement
